renamed files in accordance with Rails conventions
