+++ app/config/routes.rb
@@ -39,8 +39,8 @@ Rails.application.routes.draw do
       resource :expired_invitation, only: %i[show]
 
       # Session management
-      post "sessions/refresh", to: "sessions#refresh", as: :session_refresh
-      get "sessions/end", to: "sessions#end", as: :session_end
+      post "session/refresh", to: "sessions#refresh", as: :session_refresh
+      get "session/end", to: "sessions#end", as: :session_end
     end
 
     scope "/:client_agency_id", module: :caseworker, constraints: { client_agency_id: Regexp.union(Rails.application.config.client_agencies.client_agency_ids) } do
